refactor: 패키지 이동
view 내에 extension 패키지 생성
