+++ app/src/main/java/woowacourse/movie/presentation/view/book_complete/BookCompleteActivity.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view.book_complete
+package woowacourse.movie.presentation.view.book_complete
 
 import android.os.Bundle
 import android.widget.Toast
@@ -7,7 +7,7 @@ import woowacourse.movie.databinding.ActivityBookCompleteBinding
 import woowacourse.movie.domain.TicketBundle
 import woowacourse.movie.getParcelableCompat
 import woowacourse.movie.model.MovieBookingInfo
-import woowacourse.movie.view.common.BackButtonActivity
+import woowacourse.movie.presentation.view.common.BackButtonActivity
 
 class BookCompleteActivity : BackButtonActivity() {
     private lateinit var binding: ActivityBookCompleteBinding
