+++ app/src/main/java/woowacourse/movie/presentation/view/main/MainActivity.kt
@@ -1,9 +1,9 @@
-package woowacourse.movie.view.main
+package woowacourse.movie.presentation.view.main
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import woowacourse.movie.databinding.ActivityMainBinding
-import woowacourse.movie.view.main.data.MovieDummyData
+import woowacourse.movie.presentation.view.main.data.MovieDummyData
 
 class MainActivity : AppCompatActivity() {
     private lateinit var binding: ActivityMainBinding
