+++ app/src/main/java/woowacourse/movie/presentation/view/main/MovieListViewHolder.kt
@@ -1,9 +1,9 @@
-package woowacourse.movie.view.main
+package woowacourse.movie.presentation.view.main
 
 import android.content.Intent
 import woowacourse.movie.databinding.ItemMovieListBinding
 import woowacourse.movie.model.Movie
-import woowacourse.movie.view.movie_detail.MovieDetailActivity
+import woowacourse.movie.presentation.view.movie_detail.MovieDetailActivity
 
 class MovieListViewHolder(binding: ItemMovieListBinding) {
     private val context = binding.root.context
