+++ app/src/main/res/layout/activity_seat_picker.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.seat_pick.SeatPickerActivity">
+    tools:context=".presentation.view.seat_pick.SeatPickerActivity">
 
     <TextView
         android:id="@+id/tv_seat_picker_screen"
